refactor: convert css.js to Typescript
